+++ TrashMob/client-app/src/components/EventManagement/ManageEventSummary.tsx
@@ -1,14 +1,13 @@
 import * as React from 'react'
-import UserData from './Models/UserData';
+import UserData from '../Models/UserData';
 
-export interface ManageEventMediaProps {
+export interface ManagerEventSummaryProps {
     eventId: string;
     isUserLoaded: boolean;
     currentUser: UserData;
 }
 
-export const ManageEventMedia: React.FC<ManageEventMediaProps> = () => {
- 
+export const ManageEventSummary: React.FC<ManagerEventSummaryProps> = (props) => {
     return (
         <>
             <div>
